pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "ai-agents"
version = "0.1.0"
description = ""
authors = [
    {name = "José Cruz",email = "jose.garrido@g.globo"}
]
readme = "README.md"
requires-python = "3.13.3"
dependencies = [
    "langchain (>=1.1.0,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "black (>=25.11.0,<26.0.0)",
    "langchain-openai (>=1.1.0,<2.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "poetry-plugin-dotenv (>=3.1.0,<4.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
